backend/requirements.txt
Django>=4.2
djangorestframework
djangorestframework-simplejwt
gunicorn
redis